feat(subsciption): Monthly subscription

+++ cmd/web/render.go
@@ -28,7 +28,7 @@ var functions = template.FuncMap{
 }
 
 func formatCurrency(n int) string {
-	f := float32(n) / 100.0
+	f := float32(n) / float32(100)
 	return fmt.Sprintf("$%.2f", f)
 }
